Fold TempFiles into FolderFactory
