+++ NuKeeper/Files/FolderFactory.cs
@@ -1,4 +1,6 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
+using System.Linq;
 using NuKeeper.Logging;
 
 namespace NuKeeper.Files
@@ -12,16 +14,46 @@ namespace NuKeeper.Files
             _logger = logger;
         }
 
-        public void DeleteExistingTempDirs()
-        {
-            TempFiles.DeleteExistingTempDirs(_logger);
-        }
-
         public IFolder UniqueTemporaryFolder()
         {
-            var tempDir = new DirectoryInfo(TempFiles.GetUniqueTemporaryPath());
+            var tempDir = new DirectoryInfo(GetUniqueTemporaryPath());
             tempDir.Create();
             return new Folder(_logger, tempDir);
+        }
+
+        private static string NuKeeperTempFilesPath()
+        {
+            return Path.Combine(Path.GetTempPath(), "NuKeeper");
+        }
+
+        private static string GetUniqueTemporaryPath()
+        {
+            var uniqueName = Guid.NewGuid().ToString("N");
+            return Path.Combine(NuKeeperTempFilesPath(), uniqueName);
+        }
+
+        public void DeleteExistingTempDirs()
+        {
+            var dirInfo = new DirectoryInfo(NuKeeperTempFilesPath());
+            var dirs = dirInfo.Exists ? dirInfo.EnumerateDirectories() : Enumerable.Empty<DirectoryInfo>();
+            foreach (var dir in dirs)
+            {
+                TryDelete(dir);
+            }
+        }
+
+        private void TryDelete(DirectoryInfo tempDir)
+        {
+            _logger.Verbose($"Attempting delete of temp dir {tempDir}");
+
+            try
+            {
+                tempDir.Delete(true);
+            }
+            catch (Exception)
+            {
+                _logger.Verbose("Delete failed. Continuing");
+            }
         }
     }
 }
